Moves all vtk references to the cdat app
