+++ apps/web_fe/urls.py
@@ -3,6 +3,7 @@ from django.conf.urls import patterns, include, url
 from web_fe import views
 from velo import views as velo_views
 from esgf import views as esgf_views
+from cdat import views as cdat_views
 
 velo_patterns = [
         url(r'^get_folder/$', velo_views.get_folder),
@@ -16,6 +17,13 @@ esgf_patterns = [
         url(r'^node_info/$', esgf_views.node_info),
         url(r'^node_search/$', esgf_views.node_search),
         url(r'^load_facets/$', esgf_views.load_facets),
+]
+
+cdat_patterns = [
+        url(r'^vtk/', cdat_views.vtkweb_launcher),
+        url(r'^_refresh', cdat_views._refresh),
+        url(r'^vtk_viewer', cdat_views.vtk_viewer),
+        url(r'^vtk_test', cdat_views.vtk_test),
 ]
 
 urlpatterns = patterns('',
@@ -36,6 +44,7 @@ urlpatterns = patterns('',
                        url(r'^userdata/image/(?P<path>.*\.png)$', views.send_image),
                        url(r'^velo/', include(velo_patterns)),
                        url(r'^esgf/', include(esgf_patterns)),
+                       url(r'^cdat/', include(cdat_patterns)),
                        )
 
 urlpatterns += staticfiles_urlpatterns()
