+++ apps/web_fe/views.py
@@ -5,7 +5,6 @@ from django.template import RequestContext, loader
 from django.forms.models import model_to_dict
 from django.forms.utils import ErrorList
 from django.core.context_processors import csrf
-from django.views.decorators.csrf import csrf_exempt
 from django.core.files import File
 from forms import UserCreationForm
 from django.contrib.auth import authenticate, login, logout
@@ -15,7 +14,6 @@ from django.contrib.auth.decorators import login_required
 from django.conf import settings
 
 from web_fe.models import TileLayout, Credential
-import sys
 import json
 import simplejson
 import os
@@ -23,13 +21,9 @@ import urllib
 import urllib2
 import base64
 import time
-import datetime
 import requests
-from subprocess import Popen, PIPE
-from sets import Set
-from velo import VeloAPI
+from sendfile import sendfile
 from util.utilities import print_debug
-import vtk_launcher
 
 
 def render_template(request, template, context):
@@ -371,8 +365,6 @@ def credential_check_existance(request):
 
 @login_required
 def send_image(request, path):
-    from sendfile import sendfile
-    import os
     cred = Credential.objects.get(service='velo', site_user_name=request.user)
     fullpath = os.path.join('userdata', cred.service_user_name, path)
 
@@ -380,97 +372,3 @@ def send_image(request, path):
         return sendfile(request, fullpath)
     else:
         return HttpResponse(status=404)
-
-
-@csrf_exempt  # should probably fix this at some point
-@login_required
-def vtkweb_launcher(request):
-    """Proxy requests to the configured launcher service."""
-    import requests
-    try:
-        from django.conf import settings
-        VISUALIZATION_LAUNCHER = settings.VISUALIZATION_LAUNCHER
-    except ImportError:
-        VISUALIZATION_LAUNCHER = None
-
-    if not VISUALIZATION_LAUNCHER:
-        # unconfigured launcher
-        return HttpResponse(status=404)
-
-    # TODO: add status and delete methods
-    if request.method == 'POST':
-        req = requests.post(VISUALIZATION_LAUNCHER, request.body)
-        if req.ok:
-            return HttpResponse(req.content)
-        else:
-            return HttpResponse(status=500)
-
-    return HttpResponse(status=404)
-
-
-def print_debug(e):
-    import traceback
-    print '1', e.__doc__
-    print '2', sys.exc_info()
-    print '3', sys.exc_info()[0]
-    print '4', sys.exc_info()[1]
-    print '5', traceback.tb_lineno(sys.exc_info()[2])
-    ex_type, ex, tb = sys.exc_info()
-    print '6', traceback.print_tb(tb)
-
-
-# ******
-# VTK
-# ******
-def _refresh(request):
-    """Refresh the visualization session information."""
-    # check the session for a vtkweb instance
-    vis = request.session.get('vtkweb')
-    if vis is None or vtk_launcher.status(vis.get('id', '')) is None:
-        # open a visualization instance
-        vis = vtk_launcher.new_instance()
-        request.session['vtkweb'] = vis
-    return dict(vis)
-
-
-def vtk_viewer(request):
-    """Open the main visualizer view."""
-    data = {}
-    data['base'] = base_path
-    data['files'] = [
-            f for f in os.listdir(base_path)
-            if not os.path.isdir(os.path.join(base_path, f))
-            ]
-    data['dirs'] = [
-            f for f in os.listdir(base_path)
-            if os.path.isdir(os.path.join(base_path, f))
-            ]
-    return render(
-            request,
-            'vtk_view/cdat_viewer.html',
-            data
-            )
-
-
-def vtk_test(request, test="cone"):
-    return render(request, 'vtk_view/view_test.html', {"test": test})
-
-
-@csrf_exempt  # should probably fix this at some point
-def vtkweb_launcher(request):
-    """Proxy requests to the configured launcher service."""
-    import requests
-    VISUALIZATION_LAUNCHER = 'http://aims1.llnl.gov/vtk'
-    if getattr(settings, 'VISUALIZATION_LAUNCHER'):
-        VISUALIZATION_LAUNCHER = settings.VISUALIZATION_LAUNCHER
-    if not VISUALIZATION_LAUNCHER:
-        # unconfigured launcher
-        return HttpResponse(status=404)
-    # TODO: add status and delete methods
-    if request.method == 'POST':
-        req = requests.post(VISUALIZATION_LAUNCHER, request.body)
-        if req.ok:
-            return HttpResponse(req.content)
-        else:
-            return HttpResponse(status=500)
-    return HttpResponse(status=404)
